requirements.txt
Flask==2.2.3
torch==2.5.1
torchvision==0.20.1
tensorflow==2.16.1
Pillow==10.3.0
Werkzeug==3.0.2
ultralytics==8.3.50
